a little more easy monitor cleanup
